Manage listings Page and User Authorization

+++ app/Http/Controllers/ListingController.php
@@ -43,51 +43,70 @@ class ListingController extends Controller
             'description' => 'required'
         ]);
 
-        if($request->hasFile('logo')){
-            $formFields['logo'] = $request->file('logo')->store('logos','public');
+        if ($request->hasFile('logo')) {
+            $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
         $formFields['user_id'] = auth()->id();
 
         Listing::create($formFields);
 
-        
 
-        return redirect('/')->with('message','Listing created successfully!');
+
+        return redirect('/')->with('message', 'Listing created successfully!');
     }
 
     //Show Edit Form
-    public function edit(Listing $listing){
-        return view('listings.edit',['listing'=>$listing]);
+    public function edit(Listing $listing)
+    {
+        return view('listings.edit', ['listing' => $listing]);
     }
 
-     //Update Listing Data  
-     public function update(Request $request, Listing $listing)
-     {
-         $formFields = $request->validate([
-             'title' => 'required',
-             'company' => ['required'],
-             'location' => 'required',
-             'website' => 'required',
-             'email' => ['required', 'email'],
-             'tags' => 'required',
-             'description' => 'required'
-         ]);
- 
-         if($request->hasFile('logo')){
-             $formFields['logo'] = $request->file('logo')->store('logos','public');
-         }
- 
-        $listing->update($formFields);
- 
-         
- 
-         return back()->with('message','Listing updated successfully!');
-     }
+    //Update Listing Data  
+    public function update(Request $request, Listing $listing)
+    {
 
-     //Delete Listing
-     public function destroy(Listing $listing){
+        // Make Sure Logged In User Is Owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
+        $formFields = $request->validate([
+            'title' => 'required',
+            'company' => ['required'],
+            'location' => 'required',
+            'website' => 'required',
+            'email' => ['required', 'email'],
+            'tags' => 'required',
+            'description' => 'required'
+        ]);
+
+        if ($request->hasFile('logo')) {
+            $formFields['logo'] = $request->file('logo')->store('logos', 'public');
+        }
+
+        $listing->update($formFields);
+
+
+
+        return back()->with('message', 'Listing updated successfully!');
+    }
+
+    //Delete Listing
+    public function destroy(Listing $listing)
+    {
+
+        // Make Sure Logged In User Is Owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         $listing->delete();
-        return redirect('/')->with('message','Listing deleted successfully');
-     }
+        return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    // Manage Listings
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
+    }
 }
